requirements.txt
asgiref==3.4.1
certifi==2021.10.8
charset-normalizer==2.0.11
distlib==0.3.4
Django==4.0.2
django-bulk-sync==3.2.0
django-cors-headers==3.11.0
django-debug-toolbar==3.2.4
django-environ==0.8.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
filelock==3.4.2
idna==3.3
joblib==1.1.0
numpy==1.22.2
pandas==1.4.0
platformdirs==2.4.1
psycopg2==2.9.3
PyJWT==2.3.0
python-dateutil==2.8.2
pytz==2021.3
requests==2.27.1
six==1.16.0
soupsieve==2.3.1
sqlparse==0.4.2
urllib3==1.26.8
